bull-or-bear-backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
celery==5.4.0
redis==5.0.8
channels==4.1.0
channels-redis==4.2.0
requests==2.32.3
websockets>=13.0
daphne>=4.0
